requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2
python-jose[cryptography]
passlib[bcrypt]
streamlit
python-multipart
python-dotenv
pydantic
requests